+++ src/main/java/com/simsilica/input/MovementState.java
@@ -91,6 +91,9 @@ public class MovementState extends BaseAppState {
      *  target has been provided.
      */
     public void setMovementTarget( MovementTarget target ) {
+        if( this.target != null ) {
+            this.target.terminate(this);
+        }
         this.target = target;
         target.initialize(this);
         initializeRotation(target.getRotation());